% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{djpr_black_rock}
\alias{djpr_black_rock}
\title{Hex code for the colour: djpr black rock}
\format{
An object of class \code{character} of length 1.
}
\usage{
djpr_black_rock
}
\description{
#343641
}
\keyword{datasets}
