% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{djpr_blue}
\alias{djpr_blue}
\title{Hex code for the colour: djpr blue (main colour)}
\format{
An object of class \code{character} of length 1.
}
\usage{
djpr_blue
}
\description{
#1F1547
}
\keyword{datasets}
