% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{djpr_cobalt}
\alias{djpr_cobalt}
\title{Hex code for the colour: djpr cobalt}
\format{
An object of class \code{character} of length 1.
}
\usage{
djpr_cobalt
}
\description{
#004C97
}
\keyword{datasets}
