% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{djpr_cool_grey_1}
\alias{djpr_cool_grey_1}
\title{Hex code for the colour: djpr cool grey 1}
\format{
An object of class \code{character} of length 1.
}
\usage{
djpr_cool_grey_1
}
\description{
#d9d9d6
}
\keyword{datasets}
