% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{djpr_cool_grey_11}
\alias{djpr_cool_grey_11}
\title{Hex code for the colour: djpr cool grey 11}
\format{
An object of class \code{character} of length 1.
}
\usage{
djpr_cool_grey_11
}
\description{
#53565a
}
\keyword{datasets}
