% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{djpr_dark_pastel_green}
\alias{djpr_dark_pastel_green}
\title{Hex code for the colour: djpr dark pastel green}
\format{
An object of class \code{character} of length 1.
}
\usage{
djpr_dark_pastel_green
}
\description{
#00b140
}
\keyword{datasets}
