% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{djpr_dark_tangerine}
\alias{djpr_dark_tangerine}
\title{Hex code for the colour: djpr dark tangerine}
\format{
An object of class \code{character} of length 1.
}
\usage{
djpr_dark_tangerine
}
\description{
"#FF9E1B"
}
\keyword{datasets}
