% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{djpr_electric_lime}
\alias{djpr_electric_lime}
\title{Hex code for the colour: djpr electric lime (main colour)}
\format{
An object of class \code{character} of length 1.
}
\usage{
djpr_electric_lime
}
\description{
#cedc00
}
\keyword{datasets}
