% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{djpr_faint_blue}
\alias{djpr_faint_blue}
\title{Hex code for the colour: DJPR faint blue
"#BCD3EF"}
\format{
An object of class \code{character} of length 1.
}
\usage{
djpr_faint_blue
}
\description{
Hex code for the colour: DJPR faint blue
"#BCD3EF"
}
\keyword{datasets}
