% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{djpr_faint_green}
\alias{djpr_faint_green}
\title{Hex code for the colour: DJPR faint green
"#C0E4B5"}
\format{
An object of class \code{character} of length 1.
}
\usage{
djpr_faint_green
}
\description{
Hex code for the colour: DJPR faint green
"#C0E4B5"
}
\keyword{datasets}
