% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{djpr_golden_yellow}
\alias{djpr_golden_yellow}
\title{Hex code for the colour: djpr golden yellow}
\format{
An object of class \code{character} of length 1.
}
\usage{
djpr_golden_yellow
}
\description{
#f3e500
}
\keyword{datasets}
