% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{djpr_iris_blue}
\alias{djpr_iris_blue}
\title{Hex code for the colour: djpr iris blue}
\format{
An object of class \code{character} of length 1.
}
\usage{
djpr_iris_blue
}
\description{
#00a9e0
}
\keyword{datasets}
