% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{djpr_lima}
\alias{djpr_lima}
\title{Hex code for the colour: djpr lima}
\format{
An object of class \code{character} of length 1.
}
\usage{
djpr_lima
}
\description{
#78be20
}
\keyword{datasets}
