% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{djpr_malibu}
\alias{djpr_malibu}
\title{Hex code for the colour: djpr malibu}
\format{
An object of class \code{character} of length 1.
}
\usage{
djpr_malibu
}
\description{
#71c5e8
}
\keyword{datasets}
