% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{djpr_pacific_blue}
\alias{djpr_pacific_blue}
\title{Hex code for the colour: djpr pacific blue}
\format{
An object of class \code{character} of length 1.
}
\usage{
djpr_pacific_blue
}
\description{
#0090da
}
\keyword{datasets}
