% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djpr_pal.R
\name{djpr_pal}
\alias{djpr_pal}
\title{Create a Djpr-appropriate palette for your chart.}
\usage{
djpr_pal(n = 0, reverse = FALSE, faded = FALSE)
}
\arguments{
\item{n}{Numeric. The number of levels in your colour scale. Minimum value is
1, maximum is 10. Using more than 6 is not recommended. If you don't
specify \code{n}, a five-colour palette will be used, which may not look right.
Specify \code{n}.

By default, n = 2 will give you light orange and dark orange. Use n = "2a"
if you want light orange and cool_grey_11.}

\item{reverse}{Logical. FALSE by default. Setting to TRUE reverses the
standard colour order. Standard colour order runs from light to dark. If
you set reverse to TRUE, colours will run from dark to light.}

\item{faded}{Logical. FALSE by default. Setting to TRUE returns the faded
variations of the standard colours.}
}
\description{
Create a Djpr-appropriate palette for your chart.
}
\examples{
library(ggplot2)

p <- ggplot(mtcars, aes(x = wt, y = mpg, col = factor(cyl))) +
    geom_point() +
    theme_djpr() +
    scale_colour_manual(values = djpr_pal(n = 3))

p

# Alternatively, use djpr_colour_manual(), which is a wrapper
# around scale_colour_manual():

p <- ggplot(mtcars, aes(x = wt, y = mpg, col = factor(cyl))) +
    geom_point() +
    theme_djpr() +
    djpr_colour_manual(3)

p


# When applying colours as a fill, use djpr_fill_manual, which is
# a wrapper around scale_colour_manual:

p <- ggplot(mtcars, aes(x = mpg, fill = factor(cyl))) +
  geom_histogram() +
    theme_djpr() +
    djpr_fill_manual(3)
p


}
