% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{djpr_persian_green}
\alias{djpr_persian_green}
\title{Hex code for the colour: djpr persian green}
\format{
An object of class \code{character} of length 1.
}
\usage{
djpr_persian_green
}
\description{
#00b2a9
}
\keyword{datasets}
