% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{djpr_persimmon}
\alias{djpr_persimmon}
\title{Hex code for the colour: djpr persimmon}
\format{
An object of class \code{character} of length 1.
}
\usage{
djpr_persimmon
}
\description{
"#E35205"
}
\keyword{datasets}
