% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{djpr_purple}
\alias{djpr_purple}
\title{Hex code for the colour: DJPR purple
"#745ECF"}
\format{
An object of class \code{character} of length 1.
}
\usage{
djpr_purple
}
\description{
Hex code for the colour: DJPR purple
"#745ECF"
}
\keyword{datasets}
