% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{djpr_royal_blue}
\alias{djpr_royal_blue}
\title{Hex code for the colour: DJPR royal blue
"#2A6FA2"}
\format{
An object of class \code{character} of length 1.
}
\usage{
djpr_royal_blue
}
\description{
Hex code for the colour: DJPR royal blue
"#2A6FA2"
}
\keyword{datasets}
