% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djpr_save_image.R
\name{djpr_save_image}
\alias{djpr_save_image}
\title{Save ggplot2 plots as static images at the right size for DJPR PowerPoint decks.}
\usage{
djpr_save_image(
  filename,
  object = last_plot(),
  size = c("full", "twothirds", "half", "quarter"),
  remove_title = TRUE,
  dpi = "retina"
)
}
\arguments{
\item{filename}{name, including path and extension, of the file to be saved}

\item{object}{name of the ggplot2 plot to be saved;
default is the last plot, using \code{ggplot2::last_plot()}}

\item{size}{size of picture. Options are: "full", "twothirds",
"half" and "quarter"}

\item{remove_title}{\code{TRUE} by default. When \code{TRUE}, title and subtitle will
be removed from the plot before saving. If \code{FALSE}, title and subtitle are
retained.}

\item{dpi}{resolution of picture. Default is 'retina' which is 320 dpi.}
}
\description{
This function is a wrapper around ggplot2::ggsave(). It is designed for
saving plots as static images that will be then included in a DJPR PPTX deck.
}
\details{
Plots can be saved in pre-defined sizes: full slide, half slide,
2/3rds slide, and 1/4 slide, to fit the DJPR-SPP PowerPoint deck template.

The function removes the plot's title and subtitle by default, as these
should be added on the slide itself rather than included in the image.
}
\examples{
\dontrun{
library(ggplot2)
p <- ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point()

djpr_save_image("test.png", p)

# Or to save an image in the right size to fill half a Powerpoint slide

djpr_save_image("test.png", p, size = "half")
}
}
