% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{djpr_spray}
\alias{djpr_spray}
\title{Hex code for the colour: djpr spray}
\format{
An object of class \code{character} of length 1.
}
\usage{
djpr_spray
}
\description{
#88dbdf
}
\keyword{datasets}
