% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djpr_use_fonts.R
\name{djpr_use_fonts}
\alias{djpr_use_fonts}
\title{Enable and use the custom VIC font in plots}
\usage{
djpr_use_fonts()
}
\description{
After calling this, ggplot outputs will use the VIC font.
}
\details{
This requires the suggested packages \code{showtext} and \code{sysfonts} to work.
}
