% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{djpr_watercourse}
\alias{djpr_watercourse}
\title{Hex code for the colour: djpr watercourse}
\format{
An object of class \code{character} of length 1.
}
\usage{
djpr_watercourse
}
\description{
#00573f
}
\keyword{datasets}
