% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot_data.R
\name{get_plot_data}
\alias{get_plot_data}
\title{Extract data behind a ggplot}
\usage{
get_plot_data(plot = ggplot2::last_plot())
}
\arguments{
\item{plot}{The plot to extract data from. Currently only supports \code{ggplot}
objects. Defaults to the last ggplot modified or created.}
}
\value{
A data frame with the extracted data. Each column has been mapped
to some aesthetic in the plot.
}
\description{
Extract the data that was passed into the \code{ggplot2()} or \code{qplot()} call,
with only the columns that are mapped to aesthetics.
}
\details{
If an expression is mapped to an aesthetic, it is added as a column in
the result, with the expression as the column name. You may need to use
backticks (e.g. \verb{`column expression`}) to refer to those columns.

Mappings provided in subsequent layers are also added if those
layers don't have their own data. Layers which have their own data
(passed in as \code{geom_xxxx(data=...)}) are ignored, as these are often used to
add annotations or guides.
}
\examples{
library(ggplot2)

# Extract the data that was passed into the plot
sample_plot <- ggplot(mtcars, aes(mpg, hp)) + geom_point()
get_plot_data(sample_plot) # -> data.frame with columns: mpg, hp

# Also picks up mappings in subsequent layers
sample_plot <-
  ggplot(iris, aes(Petal.Width, Petal.Length)) +
  geom_density2d() +
  geom_point(aes(colour=Species))
get_plot_data(sample_plot)
# -> data.frame with columns: Petal.Width, Petal.Length, Species

# Supports expression mappings
sample_plot <- qplot(mpg - mean(mpg), hp^2, data=mtcars)
get_plot_data(sample_plot)
# -> data.frame with columns: `mpg - mean(mpg)`, `hp^2`

# Facet variables are included:
sample_plot <- ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point() +
  facet_wrap(~cyl)

get_plot_data(sample_plot)
# -> data.frame with columns: `wt`, `mpg`, `cyl`

}
