% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_font_change.R
\name{gg_font_change}
\alias{gg_font_change}
\title{Change the font in all text layers of a ggplot}
\usage{
gg_font_change(p, font = "Arial")
}
\arguments{
\item{p}{A ggplot2 object}

\item{font}{Name of a font, such as "Arial", "sans", or "Roboto"}
}
\value{
A ggplot object with the \code{family} parameter of text layers
set to the provided font
}
\description{
Change the font in all text layers of a ggplot
}
\examples{

library(ggplot2)
p <- ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point() +
  geom_text(aes(label = mpg))

gg_font_change(p, "serif")
}
