% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_labs.R
\name{remove_labs}
\alias{remove_labs}
\title{Remove title, subtitle, and caption from a ggplot2 object}
\usage{
remove_labs(plot)
}
\arguments{
\item{plot}{A ggplot2 object}
}
\description{
Remove title, subtitle, and caption from a ggplot2 object
}
\details{
If \code{plot} is a {patchwork} object, title, subtitle and caption
will be removed from the combined plot, but not from the sub-plots.
}
