% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_djpr.R
\name{theme_djpr}
\alias{theme_djpr}
\title{Create a ggplot2 theme consistent with the DJPR style guide.}
\usage{
theme_djpr(
  base_size = 12,
  base_family = "sans",
  chart_type = "normal",
  flipped = FALSE,
  background = "white",
  legend = "bottom",
  panel_borders = FALSE
)
}
\arguments{
\item{base_size}{Size for text elements. Defaults to 12.}

\item{base_family}{Font family for text elements. Defaults to "sans".}

\item{chart_type}{"normal" by default. Set to "scatter" for scatter plots.}

\item{flipped}{FALSE by default. Set to TRUE if using coord_flip(). If set to
TRUE, the theme will show a vertical axis line, ticks & panel grid, while
hiding the horizontals. Ignored for type = "scatter".}

\item{background}{"white" by default.}

\item{legend}{"bottom" by default. Set to "right", "left", "top" or "off" as
desired, or a two element numeric vector such as c(0.9, 0.1).}

\item{panel_borders}{\code{FALSE} by default. Set to \code{TRUE} to enable a black
border around the plotting area.}
}
\description{
Create a ggplot2 theme consistent with the DJPR style guide.
}
