% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mc.R
\name{theme_mc}
\alias{theme_mc}
\title{Style a plot in a dumb style}
\usage{
theme_mc(base_size = 16)
}
\arguments{
\item{base_size}{Base font size}
}
\value{
a ggplot2 theme object
}
\description{
Style a plot in a dumb style
}
\examples{
ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point() +
  theme_mc()
}
