% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djpr_use_fonts.R
\name{without_showtext}
\alias{without_showtext}
\title{Run an expression with showtext turned off}
\usage{
without_showtext(expr)
}
\arguments{
\item{expr}{Expression to run.}
}
\description{
This is needed because showtext interferes with slide output.
}
