% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{djpr_green}
\alias{djpr_green}
\title{Hex code for the colour: djpr green (main colour)}
\format{
An object of class \code{character} of length 1.
}
\usage{
djpr_green
}
\description{
#78be20
}
\keyword{datasets}
