% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djpr_save_image.R
\name{djpr_save_image}
\alias{djpr_save_image}
\title{Save ggplot2 plots as static images at the right size for DJPR PowerPoint decks.}
\usage{
djpr_save_image(filename, object = last_plot(), size = "full", dpi = "retina")
}
\arguments{
\item{filename}{name of the file to be saved into}

\item{object}{object of plot to be saved, default is the last plot from last_plot()}

\item{size}{size of picture. Options are: full, twothirds, half and quarter}

\item{dpi}{resolution of picture. Default is 'retina' which is 320 dpi.}
}
\description{
This function is a wrapper around ggplot2::ggsave().
}
\details{
This enables users to save images in certain pre-defined sizes: full slide, half slide,
2/3rds slide, and 1/4 slide to fit DJPR powerpoint deck template.
The function also removes title and subtitle by default, as these
should be added on the slide itself rather than included in the image.
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(x = wt, y = mpg)) +
     geom_point()

djpr_save_image("test.png", p)

# Or to save an image in the right size to fill half a Powerpoint slide

djpr_save_image("test.png", p, size = "half")

}
