% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djpr_save_pptx.R
\name{djpr_save_pptx}
\alias{djpr_save_pptx}
\title{Export a ggplot to a slide}
\usage{
djpr_save_pptx(
  plot,
  destination,
  layout = c("full", "half", "twothirds"),
  signpost = NULL
)
}
\arguments{
\item{plot}{The ggplot to insert}

\item{destination}{File path to save the slide to}

\item{layout}{Slide layout to use}

\item{signpost}{Whether there should be a signpost heading}
}
\value{
Since the function saves the slide to a file, the return value
is not defined and may change in future
}
\description{
The exported slide will be in the SPP template.
}
\examples{
\dontrun{
# Export `the_ggplot` to file `output.pptx`
djpr_save_pptx(the_ggplot, "output.pptx")
# Specify layout and signpost
djpr_save_pptx(the_ggplot, "output.pptx", layout="half", signpost="section")
}
}
