% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djpr_save_pptx.R
\name{djpr_save_pptx}
\alias{djpr_save_pptx}
\title{Export a ggplot to a DJPR SPP PowerPoint slide.}
\usage{
djpr_save_pptx(
  destination = NULL,
  plot = ggplot2::last_plot(),
  layout = c("full", "half", "twothirds"),
  signpost = NULL
)
}
\arguments{
\item{destination}{The destination for the slide. Can be one of:
\itemize{
\item \code{NULL} (default): The slide will be kept in-memory and returned for further
manipulation.
\item \code{character}: File path to save the slide to, including extension,
such as "output.pptx".
\item A \code{rpptx} object: An in-memory slide pack to append to.
}}

\item{plot}{The ggplot to turn into a slide.
Default is \code{ggplot2::last_plot()}.}

\item{layout}{Slide layout to use. Options are "full", "half", and
"twothirds". The layouts refer to the proportion of the slide width taken
up by the chart.}

\item{signpost}{Signpost heading, if required. If \code{NULL} (the default), no
signpost will be included.}
}
\value{
If \code{destination} was \code{NULL} or a \code{pptx} object, returns the updated
presentation ready for further changes.
Otherwise, the return value is undefined (and may change in future) since
the function saves the slide to a file.
}
\description{
The exported slide will use the SPP template. All text is editable on the
slide. Users can choose from several slide formats.
}
\examples{
\dontrun{
library(ggplot2)
# First, create a ggplot2 object
the_ggplot <- ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point() +
  labs(title = "Chart title",
       subtitle = "Chart subtitle",
       caption = "Notes and source information.")

# Export `the_ggplot` to file `output.pptx`
djpr_save_pptx("output.pptx", the_ggplot)

# Specify layout and signpost
djpr_save_pptx("output.pptx", the_ggplot, layout = "half", signpost = "section")

# Export multiple slides. Different plots can be passed in for each slide.
djpr_save_pptx(NULL, the_ggplot) \%>\%
  djpr_save_pptx(the_ggplot, layout="half") \%>\%
  print(target="slidepack.pptx")
}
}
